+++ src/idc/reader/depth/__init__.py
@@ -1,2 +1,3 @@
+from ._csv import CSVDepthInfoReader
 from ._grayscale import GrayscaleDepthInfoReader
 from ._numpy import NumpyDepthInfoReader