+++ src/idc/writer/depth/__init__.py
@@ -1,2 +1,3 @@
+from ._csv import CSVDepthInfoWriter
 from ._grayscale import GrayscaleDepthInfoWriter
 from ._numpy import NumpyDepthInfoWriter